refactor(actions/user): Move code to fetch user under actions.
